# Conflicts:
#	lib/model/history_card.dart
#	lib/view_model/data_view_view_model.dart
#	lib/view_model/history_card_view_model.dart

+++ lib/model/sensor.dart
@@ -1,6 +1,3 @@
-import 'dart:async';
-import 'livedata.dart';
-
 class Data{
   int timestamp;
   double mForce;
@@ -16,9 +13,9 @@ class Data{
   {
     mForce = force;
   }
-  double getTime()
+  int getTime()
   {
-    return timestamp.toDouble();
+    return timestamp;
   }
   double getForce()
   {
